LayoutTests/platform/wpe/media/track/track-in-band-legacy-api-expected.txt


Test that in-band tracks work with .webkitClosedCaptionsVisible.

EVENT(canplaythrough)

** Check initial in-band track states
EXPECTED (video.textTracks.length == '1'), OBSERVED '0' FAIL
RUN(inbandTrack1 = video.textTracks[0])
TypeError: undefined is not an object (evaluating 'inbandTrack1.mode')
TypeError: undefined is not an object (evaluating 'inbandTrack1.cues')
TypeError: undefined is not an object (evaluating 'inbandTrack1.language')
TypeError: undefined is not an object (evaluating 'inbandTrack1.kind')

** Enable cues and let them load
RUN(video.webkitClosedCaptionsVisible = true)
RUN(video.play())
EVENT(seeked)

** Test to make sure captions are displaying
TypeError: undefined is not an object (evaluating 'inbandTrack1.cues')

** Disable captions, cues should be flushed
RUN(video.webkitClosedCaptionsVisible = false)
TypeError: undefined is not an object (evaluating 'inbandTrack1.cues')

END OF TEST
